workflow-orchestrator/requirements.txt
docker>=6.0.0
pyyaml>=6.0
pathlib2>=2.3.7; python_version < "3.4"
requests>=2.28.0
psutil>=5.9.0
